prepare KokuraexWeb.UtilFunction & move function is_eq_or_gt_atom/1 in there
