+++ app/lib/kokuraex_web/controllers/event_functions/event_function.ex
@@ -4,6 +4,7 @@ defmodule KokuraexWeb.EventFunction do
   """
 
   import KokuraexWeb.DatetimeFunction
+  import KokuraexWeb.UtilFunction
 
   @doc """
   GET event data from connmass API.
