+++ app/lib/kokuraex_web/controllers/util_functions/util_function.ex
@@ -1,0 +1,12 @@
+defmodule KokuraexWeb.UtilFunction do
+  @moduledoc """
+  Utility functions.
+  """
+
+  @doc """
+  :eqか:gtならtrue、:ltならfalseを返す
+  """
+  def is_eq_or_gt_atom(:eq), do: true
+  def is_eq_or_gt_atom(:gt), do: true
+  def is_eq_or_gt_atom(:lt), do: false
+end
